backend/rag_scripts/requirements.txt
pinecone
langchain
langchain-community
langchain-huggingface
langchain-pinecone
sentence-transformers
pypdf
python-dotenv
